MODIFIED: MazeCreator
Cuidado que hemos cambiado muchas cosas.
Creemos que funciona (?)
