+++ TheHallwaySol/Project/TheHallway/TheHallway.vcxproj
@@ -161,15 +161,15 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\Src\TheHallway\Player\Lighter.cpp" />
-    <ClCompile Include="..\..\Src\TheHallway\Player\PlayerMovement.cpp" />
+    <ClCompile Include="..\..\Src\TheHallway\Lighter.cpp" />
+    <ClCompile Include="..\..\Src\TheHallway\MazeCreator.cpp" />
+    <ClCompile Include="..\..\Src\TheHallway\PlayerMovement.cpp" />
     <ClCompile Include="..\..\Src\TheHallway\Source.cpp" />
-    <ClCompile Include="MazeCreator.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\Src\TheHallway\Player\Lighter.h" />
-    <ClInclude Include="..\..\Src\TheHallway\Player\PlayerMovement.h" />
-    <ClInclude Include="MazeCreator.h" />
+    <ClInclude Include="..\..\Src\TheHallway\Lighter.h" />
+    <ClInclude Include="..\..\Src\TheHallway\MazeCreator.h" />
+    <ClInclude Include="..\..\Src\TheHallway\PlayerMovement.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
