+++ TheHallwaySol/Project/TheHallway/TheHallway.vcxproj.filters
@@ -24,25 +24,25 @@
     <ClCompile Include="..\..\Src\TheHallway\Source.cpp">
       <Filter>Archivos de origen</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Src\TheHallway\Player\Lighter.cpp">
-      <Filter>Archivos de origen\Player</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\Src\TheHallway\Player\PlayerMovement.cpp">
-      <Filter>Archivos de origen\Player</Filter>
-    </ClCompile>
-    <ClCompile Include="MazeCreator.cpp">
+    <ClCompile Include="..\..\Src\TheHallway\MazeCreator.cpp">
       <Filter>Archivos de origen</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\Src\TheHallway\PlayerMovement.cpp">
+      <Filter>Archivos de origen\Player</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\Src\TheHallway\Lighter.cpp">
+      <Filter>Archivos de origen\Player</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\Src\TheHallway\Player\Lighter.h">
-      <Filter>Archivos de encabezado\Player</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\Src\TheHallway\Player\PlayerMovement.h">
-      <Filter>Archivos de encabezado\Player</Filter>
-    </ClInclude>
-    <ClInclude Include="MazeCreator.h">
+    <ClInclude Include="..\..\Src\TheHallway\MazeCreator.h">
       <Filter>Archivos de encabezado</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Src\TheHallway\Lighter.h">
+      <Filter>Archivos de encabezado\Player</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Src\TheHallway\PlayerMovement.h">
+      <Filter>Archivos de encabezado\Player</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
