+++ TheHallwaySol/Src/TheHallway/Lighter.cpp
