+++ TheHallwaySol/Src/TheHallway/Lighter.h
@@ -12,6 +12,7 @@ public:
 	~Lighter() {}
 
 	virtual bool init(luabridge::LuaRef parameterTable = { nullptr })override;
+	static std::string GetName() { return "Lighter"; }
 	void charge();
 	void update() override;
 };
