+++ TheHallwaySol/Src/TheHallway/PlayerMovement.h
@@ -13,6 +13,7 @@ public:
 	~PlayerMovement() {}
 
 	virtual bool init(luabridge::LuaRef parameterTable = { nullptr }) override;
+	static std::string GetName() { return "PlayerMovement"; }
 
 	void move();
 	void update() override;
